+++ packages/pure-parse/src/internals/index.ts
@@ -1,3 +1,4 @@
 export * from './hasKey'
 export * from './Equals'
 export * from './utility-types'
+export * from './optionalSymbol'
