+++ packages/pure-parse/src/internals/optionalSymbol.ts
@@ -1,0 +1,4 @@
+/**
+ * Used to represent optional guards at runtime and compile-time in two different ways
+ */
+export const optionalSymbol = Symbol('optional')
