+++ packages/pure-parse/src/parse/index.ts
@@ -1,7 +1,6 @@
-export * from './parse'
+export * from './types'
 export * from './array'
 export * from './object'
 export * from './union'
 export * from './primitives'
 export * from './fallback'
-// Do not export optionalSymbol
