--- packages/pure-parse/src/parse/optionalSymbol.ts
@@ -1,4 +1,0 @@
-/**
- * Used to represent optional parsers at runtime and compile-time in two different ways
- */
-export const optionalSymbol = Symbol('optional parser')
